home_node/backend/private/services/api/requirements.txt
asyncio-connection-pool
asyncssh
bcrypt
litestar
msgspec
orjson
pyjwt
pyopenssl
uvicorn[standard]
websockets
python-jose[cryptography]
cryptography
